Rename the node module runner
